ras/base: effectively revert d08be7457
With the prrte model, the logic to allow for launching
an orted on a node along with mpirun is no longer applicable,
as prun doesn't play the same role as mpirun does in the orted
imbedded in Open MPI model.

Further the option was breaking prun launches on Cray using SLURM.

So, this PR removes launch_orted_on_hn and the associated
MCA parameter.  prun launches on Cray systems now allow
for stdout/stderr to returned to prun.

An attempt to revert d08be7457 which introduced the launch_orted_on_hn
didn't work, hence the 'by hand' removal in this commit.

Signed-off-by: Howard Pritchard <hppritcha@gmail.com>