Few missing prrte prefixes and fix -hostfile/-host
Couple of places were missed when prefixing MCA component structs.

Collect all hostfile and -host options and pass them to "prte". Currently passing the hostfile as a comma-separated list of filenames - need to check and see if that is correctly handled

Fix problem with matching hostnames (e.g., "localhost" to the actual "hostname")

Signed-off-by: Ralph Castain <rhc@pmix.org>