Update report of bindings
Provide --report-bindings output as per @markalle's description. This
ports the functionality in the cited PR to PRRTE, modified to conform
to OMPI/PRRTE coding standards. I also eliminated having two different
ways of reporting the bindings so we don't confuse users, and extended
the original work to support HWLOC 1.x as well as 2.x.

Note that the topology tree always includes the "complete" cpuset for
each object along with the allowed cpuset (in the "cpuset" field). So
printing the binding map with all processors shown simply requires using
the "complete" cpuset field - you don't need to collect the
"whole_system" topology.

For detection purposes, pass the "signature" containing the list form of
both the complete and cpuset fields. Only include the complete field if
the two differ - otherwise, just leave that field blank. We don't
actually use the signature for anything other than detecting that there
is a difference between two machines, so this will suffice - if two
machines have different cgroups, then we will treat the topologies as
different.

Refs https://github.com/open-mpi/ompi/pull/6760

Signed-off-by: Ralph Castain <rhc@pmix.org>

dd

Signed-off-by: Ralph Castain <rhc@pmix.org>