Not finding a job object isn't necessarily an error
If the odls gets ready to launch a job, but that job object has
disappeared, then it is safe to assume that it was removed due to a
command. It almost certainly isn't an error, and no matter what, the DVM
isn't allowed to simply die because a job might have a problem.

Signed-off-by: Ralph Castain <rhc@pmix.org>