Support user-directed refresh of dmodex cache
If the user requests a refresh of information about a remote process,
provide a mechanism by which that can be supported. Convert the dmodex
code to use PMIx pack/unpack functions so the dmodex request attributes
can be passed to the remote daemon for handling.

Also fix a segfault that can occur in prte if one of the daemons fails
and prte is shutting down.

Signed-off-by: Ralph Castain <rhc@pmix.org>