Protect against large local jobid's
We seem to be using the local jobid in an array somewhere as we are
segfaulting once we cross 32k even though the local jobid is uint16.
Regardless, once we get to the 32k point, begin cycling across jobids
to find the first unused one and reassign it.

This means we really slow down once we get past 32k jobs, and we might
hang if someone runs more than 32k jobs at a time. Need something a
little more robust here, but this fixes the problem for the moment.

Signed-off-by: Ralph Castain <rhc@pmix.org>