Protect against DVM collisions
Ensure the DVM is unique by appending its pid to the "dvm" job-level
session directory name. Add a new entry point to "schizo" to detect prun
operating as a proxy so we aren't solely depending upon the name
"mpirun". Adjust schizo component priorities so we detect the first
proxy and stop.

Signed-off-by: Ralph Castain <rhc@pmix.org>