Deal with deprecated cmd line options and rndz files
Detect if we are launching as a proxy for "prun" and will therefore be
starting our own "prte" - if so, then define a rendezvous file to ensure
that prun connects to the correct prte.

First cut at a way of dealing with deprecated CLI - in this case, deal
with the --oversubscribe option as that is causing numerous MTT
failures.

Signed-off-by: Ralph Castain <rhc@pmix.org>