support for cmdline "--entry lib"
A schizo_entry directory registers a couple command line options:
  --entry lib
  --entrybase lib
and also allows the syntaxes
  mpirun --mca ompi_tools_entry lib
  env OMPI_MCA_ompi_tools_entry=lib mpirun ...

The resulting setting is made visible to OMPI in an OMPI MCA as
ompi_tools_entry, and a wrapper is prepended in front of the app being
launched.  The wrapper is there so LD_LIBRARY_PATH and LD_PRELOAD
can be modified before the app starts.

This supports a larger checkin in OMPI where a new library
intercepts all the MPI calls and allows dynamic layering of
PMPI wrappers.

Signed-off-by: Mark Allen <markalle@us.ibm.com>