Begin conversion to use of nspace
Eventually, we want to replace the integer jobid with the PMIx nspace.
This is necessary for dealing with 3rd-party tools that will present us
with arbitrary namespaces, which we are already beginning to encounter.
This begins the process by maintaining both the original uint32_t jobid
and the corresponding nspace.

Signed-off-by: Ralph Castain <rhc@pmix.org>