remove extra [ in cset2mapstr (--bind-to :REPORT output)
There was an extra [ being printed showing hosts as
  [[<../..>][<../..>]
for example instead of
  [<../..>][<../..>]

Also there are some loops that need to iterate at least once regardless
of whether a socket or a core object is found, and I agree with whoever
started to change those from "while" to "do while".  I had initially used
a "fake_on_first_socket" and "fake_on_first_core" variable, and somebody
partially changed it to "do while" which makes more sense for this
usage.  So I completed the change to "do while" and removed the "fake_*"
vars completely.

Signed-off-by: Mark Allen <markalle@us.ibm.com>