hwloc_base_set_binding_policy using jdata too early
Without this fix the command
    mpirun --bind-to core:REPORT -np 2 hostname
segv's because hwloc_base_set_binding_policy() is called early before
jdata is set.  This fix just declines to try setting an attribute
in jdata if jdata is NULL.  The hwloc_base_set_binding_policy()
call happens again later when jdata is set, so the REPORT instruction
is still seen.

Signed-off-by: Mark Allen <markalle@us.ibm.com>