PE-LIST should't force BIND_TO_CPUSET if a binding is otherwise specified
If PE-LIST is given then the binding was being set to PRRTE_BIND_TO_CPUSET,
but I think that should only happen if a binding wasn't otherwise
specified.  If a binding was given, it can still be used in the presence
of PE-LIST, it's just that PE-LIST functions equivalently to a cgroup in
that case, constraining what PU's are allowed.

This PR uses hwloc_topology_restrict() to make the PE-LIST path work
exactly like the cgroup case, if a binding was specified.  If a binding
wasn't specified then it still lets the binding be set to
PRRTE_BIND_TO_CPUSET like the code was doing before.

There was also a too-aggressive error check in prrte_rmaps_rank_file_register()
that was active not only in the case of a rankfile being used, but was also
active if
    PRRTE_BIND_TO_CPUSET == PRRTE_GET_BINDING_POLICY(prrte_hwloc_binding_policy) &&
    !PRRTE_BIND_ORDERED_REQUESTED(prrte_hwloc_binding_policy)
which was causing any use of BIND_TO_CPUSET to fail (eg the case of not
specifying a binding but using PE-LIST, where BIND_TO_CPUSET gets set
as the binding).

Example failing command without the change to the above error check:
% mpirun --np 2 --map-by core:,PE-LIST=4-99 ./pp.x
> Conflicting directives for mapping policy are causing the policy
> to be redefined:
>   New policy:   RANK_FILE
>   Prior policy:  BYCORE

Signed-off-by: Mark Allen <markalle@us.ibm.com>