connect cmdlines to MCA settings
There's a disconnect between command line options and MCA settings in
two ways:

1. settings are queried through completely different mechanisms in
   different parts of the code:
   a. prrte_cmd_line_get_param(prrte_cmd_line, "bind-to",,)
   b. prrte_mca_base_var_register("prrte", "hwloc", "base", "binding_policy",)
   those settings diverge depending on how the bind-to option is specified.
2. there's an ordering issue at least in hwloc where initialization happens
   very early so even if --prtemca hwloc_base_binding_policy .. is on the
   cmdline, prrte_mca_base_var_register() won't see it because it happens
   too early where the cmdline hasn't been processed yet.

An extension of the same problem is --bind-to something:REPORT where the
parsing of the --bind-to argument creates a PMIX_BINDTO setting that controls
whether reporting happens, but PRRTE_MCA_hwloc_base_report_bindings is a
step further removed, because the prte.c cmdline parsing that processes
"bind-to" is separate from the later code that looks at its value, and
that's the code that needs to see :REPORT

This checkin has three locations in prte.c where it adds creation of MCA
settings based on cmdlines, and modifies the processing of command lines
based on MCA settings.

- observe the cmdline for options like --bind-to .. and
  --prtemca hwloc_base_binding_policy .. and set an equivalent MCA
  such as PRRTE_MCA_hwloc_base_binding_policy

  This is done extra early for some MCA settings because they will be
  read during rte_init() before the cmdline is looked at otherwise,
  for example hwloc init happens in rte_init() and it looks at quite
  a few MCAs like the binding policy, and mapping policy's PE-LIST

- a little later do the same thing for the rest of the cmdline options
  that didn't need to be handled extra early.  A second purpose of
  this section is that the cmdline could have additions by schizo
  for example, so we need to re-examine it anyway

- when the cmdline options are processed with prrte_cmd_line_get_param()
  I've added a prrte_cmd_line_get_param_or_env() that lets some of
  them notice if there's an MCA setting filling the same role.
  In that area there's also one case that's more complex, where the
  "bind-to" option is being looked at, and not only does it look for
  "PRRTE_MCA_hwloc_base_binding_policy" but it also looks for
  "PRRTE_MCA_hwloc_base_report_bindings" and if it finds that then
  it adds :REPORT to the existing setting (or makes one if there isn't
  an existing setting).

Related to the above, in prted.c a small portion of the above is done.
There the command lines are simpler, with just things like
    --prtemca hwloc_base_binding_policy stuff
which will eventually go into the environment and be visible to
prrte_mca_base_var_register(), but like before hwloc for example
is intiailized too early so we need to scan the cmdline early for
some of those options.

Signed-off-by: Mark Allen <markalle@us.ibm.com>