Update handling of map/rank/bind specifications
This PR addresses the rather broad issue of clearly differentiating
between setting default values using MCA parameters versus setting
per-job parameters on the cmd line or in a comm_spawn request. Given
that PRRTE is a persistent DVM (which OMPI repurposes as a one-shot RTE
when executed as "mpirun"), it is critical that the MCA parameters only
set default policies - specifically, one cannot convert a cmd line
option into an MCA parameter for the simple reason that the MCA
parameters are only read at initial daemon startup.

In the course of making that change, a number of other issues were
uncovered and resolved:

* Fix --nolocal operation
* convert SOCKET --> PACKAGE to track changed terminology in the
  overall HPC community and within HWLOC
* Fix prte_info and mapping options such as PE-LIST and PE=N
* Correctly handle error if cannot bind to specified cpus/rank
* Remove the "map" binding output as it isn't as clear/simple
  as the string output
* Add some job-description info to the non-devel job map
  Fix backward compatibility and a bug in remote topo handling
  Get the system to work with HWLOC v1.x and PMIx v3.x by protecting the
  areas that require newer versions. Fix one place where the topology
  summary wasn't being setup if the vpid=1 daemon's topology was different
  from that of the DVM master's.
* Ensure the summaries get generated when hetero topologies are sent
  to compute node daemons.

Signed-off-by: Ralph Castain <rhc@pmix.org>