Fix the daemon name issue
When assigned a name by an external body (e.g., when fork/exec'd by PMIx
or directly by a tool), the format of the name is outside of PRRTE's
control. In particular, it could contain a '.' in it, which is what
PRRTE uses to delimit the job family from the local jobid when
converting to/from string.

We know when we are registering the job family, so take advantage of
that knowledge to avoid looking for the '.'. When we are converting the
jobid of one of our children, then search the nspace backwards for the
first '.' and use that as the delimiter.

Signed-off-by: Ralph Castain <rhc@pmix.org>