Rename all non-documentation instances of prrte to prte
Unify the name throughout the code base: prte.

In documentation, the formal name of this project is the PMIx
Reference Run Time Environment (PRRTE).  But beause users and
developers alike do not care for typing two "r"s in a row, the entire
code base uses the name "prte" for:

- API symbols
- executables
- CLI params
- environment variables
- filenames
- ...everything

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>