Don't require pandoc on developer builds
Developers don't typically refer to man pages, and requiring pandoc on
every developer's machine is proving rather controversial. So if we
detect we are building from within a git clone, only require pandoc and
build the manpages if the developer requests it. Otherwise, build the
manpages if there is an adequate pandoc available, but don't if not.

Signed-off-by: Ralph Castain <rhc@pmix.org>