Add contrib recipes for Dockerizing PRRTE
Provide recipes for opensuse/leap15, centos7, and centos8. Create a
container that holds the required support libraries (including libevent,
hwloc, and libjansson) at specified revision levels. Clone and build the
PMIx and PRRTE master branches. Setup the container to start "prte" and
let it run. The rendezvous points for any required shared files can be
provided in a mounted volume.

Signed-off-by: Ralph Castain <rhc@pmix.org>