ALPS RAS: add COBALT_JOBID to list of vars
to search for to get the batch scheduler's job id.
This job id shows up in the apsched -r output, which can then
be groked to get the alps resid, which in turns gives prte
the reservation information.

Related to #577

double checked on a alps/pbspro system.
prte and prun run nominally on the alps/pbspro system.

Signed-off-by: Howard Pritchard <hppritcha@gmail.com>