Start work on tracking available vs allocated slots
We no longer have the option of modifying the base topology to adjust
for things like -host assignment of slots. This must now be done on a
per-job basis. We introduced the "slots_available" field to the
prte_node_t structure, so now begin to use it to track the number of
slots being made available to the job being mapped. This must all be
done within the mapping operation to ensure that the mappers "see" the
same value during a mapping operation as the value will be reset between
jobs.

Signed-off-by: Ralph Castain <rhc@pmix.org>