ALPS: fix an issue with using setsid with ALPS PLM
Turns out Cray ALPS can be configured such that running aprun
in a different session than the one a user is initially in when
placed on a mom node can cause aprun not to see the reservation
associated with the job.

Disable use of setsid with the --daemonize option when in a
Cray ALPS environment.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>