Ensure FT is off by default
Add the --enable-prte-ft configure option that defaults to "disable". Users
wanting to explore the new FT capabilities of PRRTE can enable this
option. We may be able to remove it (or default to enabled) when some of
the bugs have been worked out.

Move some component-specific variables out of the global area and into
that component. They are only used in that one location.

Set a global prte_enable_ft MCA param so that those who configure with
FT support can dynamically turn it on/off for experimentation.

Signed-off-by: Ralph Castain <rhc@pmix.org>