Support indirect launch by debuggers
When prun is started under a debugger, it must act as a relay for spawn
and other calls that have to be serviced by prte. There are two
use-cases that have to be considered in that mode:

* the debugger doesn't understand the command line and just passes it
  along to prun for processing. In this case, the launch directives
  impacting prun's behavior must be passed in the job_info.

* the debugger parses its command line. In this case, there are two
  spawn commands issued - the first to start prun without any cmd line
  arguments, and the second to spawn the application.

This commit resolves the first option. Another commit will follow to
address the second case.

Signed-off-by: Ralph Castain <rhc@pmix.org>