Continue work on tool support
Protect the daemons from attempting to process local events prior to actually being ready to do so. For example, a debugger could spawn "prterun", which causes PMIx_server_init to pause until the debugger has connected to it. The debugger then releases prterun to continue launching the application. However, that release event gets cached, and so prterun gets hit with it before it can complete prte_init. This leads to a segfault as prte isn't ready yet to process anything.

So protect all that by simply adding a check for "initialized" and ignoring any events until that has been set.

Signed-off-by: Ralph Castain <rhc@pmix.org>