Fix assignment of slots to apps when no -np provided
We were mistakenly computing and assigning the number of procs for
each app-context while attempting to determine default mapping/binding
policies for the job. We really only needed to know if there were
less than or more than 2 procs. Leave the computation of actual procs
to spawn when no -np is provided to the mappers.

Thanks to Michael Karo for reporting the problem.

Signed-off-by: Ralph Castain <rhc@pmix.org>