Cleanup topology sharing
We only want to communicate topologies to compute node daemons
if we are operating in a heterogeneous situation - i.e., where
nodes that might house application procs differ in some way
from each other. We weren't doing a very good job of inferring
whether or not we were in that situation, so instead just
directly detect in when we get the topology signature back from
each compute node.

Thanks to @acolinisi for the initial report and patch

Signed-off-by: Ralph Castain <rhc@pmix.org>