Try to do a better job of matching topos and nodes
Don't attempt to force the topo array to match on each
system - this can lead to errors as each node starts out
by putting its topology in index zero. Instead, do the
assignment based on signature, using the provided index
to find the corresponding signature and then identifying
the topology from that. This will only impact scalability
if there are a large number of topologies in the system,
which would be a very odd cluster indeed!

Index the nodes by their daemon's rank and provide the
corresponding topology sig index so the correlation
between the two is direct and clear.

Signed-off-by: Ralph Castain <rhc@pmix.org>