Limit --default-hostfile to prterun and prte
This specifies the total set of machines available which is useful for
unmanaged environments. The --hostfile/--host options can choose a
subset from this larger set. This option is meant to be global in
nature, and set by an admin or whoever controls the job launch in the
allocation.

Ref: https://github.com/openpmix/prrte/issues/709

Signed-off-by: Raghu Raja <craghun@amazon.com>
(cherry picked from commit ad55b396d71397ba060252e6585fd4b4366e1688)