Enable singleton "comm spawn" operations
Detect that we have been started by a singleton and
self-register the job. Provide a keepalive pipe
option so we can suicide if the singleton dies.
Return a PMIx error constant when something goes
wrong in spawn. Ensure we don't leave the client
hanging if there is an error in connect.

Signed-off-by: Ralph Castain <rhc@pmix.org>