Provide runtime warning if unable to compress
If PRRTE is built against an early version of PMIx that does
not provide compression support, and if PRRTE itself is unable
to find an adequate compression library, then provide a one-time
warning. Toggle it off with an MCA param.

If PRRTE is able to find adequate compression support, provide
an MCA param to control the minimum message size before we
start to compress.

Signed-off-by: Ralph Castain <rhc@pmix.org>