build: Fix handling of prte_get_version.sh
Two semi-related fixes for prte_get_version.sh.  First,
prte_get_version.sh is committed to Git, so it should not be
removed as part of maintainer-clean.  Second, because autogen.pl
is no longer generated from prte_get_version.m4sh, it needs to
be included in the dist tarball so that autogen.pl can be run
on a dist tarball.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>