Extend support for PMIx_Connect and fix Fence problem
PMIx_Connect requires that we perform a "fence" across all
the participants. At the conclusion of the "fence", all participants
are supposed to have access to not only the job-level info
from all participating nspace's, but also the endpoint info "put"
by all participants. Add the fence and seed it with any "put"
info.

In addition, add the generation of a context ID on behalf of
the participants so they can use it (if they choose) as, for
example, a new communicator ID and avoid multiple collective
operations. Number the assigned values from the max ID down
so we only hit a conflict if someone tries to create a massive
number of new communicators...which they shouldn't do.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 34e525cea6cfd2cefaeea4bbee90be6096c3ba62)