build: Change default build mode for components
Default to bilding MCA components into the library, rather than
as dso objects, since most users are not taking advantage of
per-library packaging, and this results in less stat/open/etc.
calls during application startup.

Clean up the static/dso selection logic for componets at the same
time.  Before, it was relatively hard to describe what would happen
if there was a mix of --enable-mca-dso and --enable-mca-static
arguments.  Now, the code will first look for a component-specific
argument, then a framework-level argument, then a global argument.
If there is a tie, static is preferred.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>