Properly handle the "inherit" directive
When used with "mpirun", users expect the initial mapping
policy to be the default and for that to be inherited by
any child jobs unless told otherwise. Push the directives
into the environment to make that happen. Fix a bug in
rmaps base where it didn't save the default policy.

Signed-off-by: Ralph Castain <rhc@pmix.org>