build: Fix compiler attribute detection
Move the #include <stdlib.h> from the body to the prologue
of AC_LANG_PROGRAM, so that it isn't included inside of
main().  Also use AC_LANG_SOURCE for the individual attribute
test, as many of them are checking things that must be
declared outside of main().  With this, modern GCC finds
all attributes checked available, as expected.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>