atomics: Only support C11 and GCC builtin atomics
Remove the inline assembly and hand-crafted assembly version
of the atomics used in PRRTE, always relying on either the
GCC-style builtin or C11 atomic operations.  All compilers
PRRTE supports include support for one of the two, even if
at a slightly less than optimal performance, and this allows
for a big code removal.

Also remove the high resolution timer code, as it was unused
in PRRTE.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>