Revise the hotel class to spread out occupants
Instead of returning a vacated room to the head of the
list for checkin, cycle across the entire hotel so
occupants aren't constantly reusing the same room.

The algorithm is a little slower, but in reality the
way we use it means that we will almost always just
move to the next room, which is quick.

Signed-off-by: Ralph Castain <rhc@pmix.org>