Check the datatype on debug "stop" attributes
Currently support "bool" by equating it to a rank of
WILDCARD and pmix_rank_t for either a specific rank
or "all" (a.k.a., WILDCARD). Return a not-supported
error for a pmix_data_array_t or any other type - we
may eventually support the array, but not now.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 892d957646465b8d0c3a870efcdda87ca54d213f)