Complete update of indirect example use-case
Indirect should limit its registration of READY_FOR_DEBUG
to just the job in which it is interested as there could
be multiple jobs in PRRTE. `prun` needs to register for
the READY_FOR_DEBUG event on the app it launches so that
the event can be relayed to the debugger. Since the debugger
is fork/exec'ing the `prun` tool, enable `prun` to get its
process ID from the environment so that it and the debugger
agree on `prun`'s name.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit b5a4d5bf077ebbe16fdf2acc05ea4a1f9ba47bdf)