Count contributors instead of children
When computing the number of contributions a daemon should
expect to receive for a given collective operation (e.g.,
fence), calculate the number of children that have a
participating daemon underneath them. This extends the search
to the bottom of the routing tree instead of just to the
child itself.

Also silence a couple of warnings.

Signed-off-by: Ralph Castain <rhc@pmix.org>