Be a little more aggressive on abnormal term
Don't look at timing of ctrl-c inputs as it gets annoying to
try and hit them fast enough. If someone gives us three ctrl-c
hits, then just die. Don't use the odls to do the final kill
as it loops thru the event library, which may be stuck - just
hammer the children and forget it.

Signed-off-by: Ralph Castain <rhc@pmix.org>