Temporary fix for prun delay
When given a dvm-uri option, prun is not correctly executing
the delay-and-retry procedure to find the file. Passing the
wait-for-connect option also has no impact.

Fix temporarily by doing the pause in prun itself.

Also noted that fence doesn't seem to be circulating info
as it should. Cleaned up a minor potential issue.

Signed-off-by: Ralph Castain <rhc@pmix.org>