Deprecating --debug option in mpirun and prte personalities.
the --debug option will print a warning that it will be deprecated starting in
Open MPI v5 in the mpirun personality. It will also warn the user that it will
be deprecated in favor of --debug-devel in the prte personality.

Signed-off-by: Nikola Dancejic <dancejic@amazon.com>
(cherry picked from commit de81a8166aef3f38c11f1226b43b0909bc5005c1)