Cleanup some warnings/format and fix simple_spawn test
The simple_spawn test was incorrectly having every initial
job process call "spawn" instead of just rank=0. This created
a confusion of new child processes. In addition, each process
that did call "spawn" was the only one who knew the nspace
of the child job - yet a "connect" was called across all members
of the parent job. This caused "connect" to hang.

Signed-off-by: Ralph Castain <rhc@pmix.org>