Do not resolve managed allocations, handle FQDN
Restore historical behavior of not resolving managed allocations
to avoid lengthy startup delays. If we are not retaining FQDN,
then ensure that the "official" hostname is the non-FQDN version
while retaining the FQDN version as an alias. If we are retaining
FQDN, then reverse the two.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit d2c6ef63131219b195dd427846b1a216bc93277c)