Add initial support for the Hydra launcher
Add a new "hydra" personality to the schizo framework with
initial implementation of a translator from hydra cmd line
syntax to PRRTE internals. This seems to work better than
trying to create an entirely new cmd line as the code in
prte and prun looks at the cmd line directly when interpreting
it.

Raises the question: should we just define a single cmd line
and use translators for alternatives? Seems like it would be
the right way to go based on this experience.

Signed-off-by: Ralph Castain <rhc@pmix.org>