prte.c: remove an unnecessary call to free()
This patch remove a call to free() in main, because the free()
is called on the return of getenv("MPIEXEC_TIMEOUT").
According to document, the return value of getenv() cannot be
modified().

To make the code clear, this patch also introduced a new
variable "timeoutenv", to save the return value of getenv()
(replacing "param").

This is because in main(), "param" is always used to stored
a pointer that need to be freed.

Signed-off-by: Wei Zhang <wzam@amazon.com>
(cherry picked from commit 9dcce32f584aa68bc15fa02ad70327b81826d0ac)