autogen.pl: ignore all excluded components
If multiple components in a framework were excluded, then only the
first one in the list was being checked - which means that only the
first one in the list was actually being excluded. All others were
missed by the current code. This change actually searches the entire
list of excluded components to see if the one we found matches any
of the list members.

Port of https://github.com/open-mpi/ompi/pull/9040

Signed-off-by: Ralph Castain <rhc@pmix.org>