Ensure we die if daemons cannot connect back
If we get an error in the daemon callback function, then
we need to ensure that we die. We cannot rely on termination
commands successfully reaching daemons as we may not know
how to reach them, so just exit and rely on the suicide
procedures of the daemons to cleanup.

Signed-off-by: Ralph Castain <rhc@pmix.org>