Redesign the command line system to provide isolation
As the number of organizations using PRRTE grows, we are
beginning to see conflicts surrounding the command line
interface. This is the primary interaction point to the
user, so it is natural that organizations want to customize
it - i.e., to "rebrand" PRRTE to match their needs.

The original code utilized a custom command line parser
that was rigidly tied to the needs of OpenMPI. Modifying
this to create the desired independence would be challenging.
Instead, we chose to replace that code with use of the
standard "getopt_long" function. This reduced the amount
of custom code, though it still requires one to properly
deal with the input/output to/from that parser.

This represents an initial working prototype of the revised
system. Note that it does change the schizo interface. Some
minor configuration changes have also crept into this PR
and will be separated out prior to commit.

Main purpose of this PR is to provide the community an
opportunity to look over the changes and provide feedback.
Updates to the code will continue. In particular, we need
to seperate out the map/rank/bind policy assignments so
that the individual organizations can each choose their
own defaults.

Signed-off-by: Ralph Castain <rhc@pmix.org>