Revamp the map-by NUMA support
For each unique topology in the system, compute the max os_index
of the CPU NUMAs by searching for the first instance of an
overlapping NUMA. We assume that any overlap stems from a GPU
NUMA, and that the os_index of such NUMAs starts at 255 and counts
downward. Cache that cutoff and use it when computing number of
NUMA objects and retrieving the Nth NUMA object.

Need to extend this to the distance computations and a few other
areas, but this covers the typical range of use-cases.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit d23ac62339ab2435c30ac8c83c33ce4bcdfb73c8)