Increase speed of rank computation
If the mapping and ranking objects are the same, then
we know that the procs were placed in the node array
in object order. This allows a shortcut when computing
their ranks - take advantage of it.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit fe89cee2770e473c075d4a8453753e4b7b1e4b78)