Enable binding internal PRRTE progress thread
In some scenarios, it is desirable to bind progress threads
to one or more specific cores - e.g., a core designated for
utility usage and not assigned application processes. Provide
an attribute by which the user can specify the core(s) to
which the internal PRRTE progress thread is to be bound, and
to indicate if the binding is mandatory (i.e., return an
error if the thread cannot be bound).

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 0828725dcde4ceb72380191eab2c7c31d4be776b)