Cleanup handling of allocated node names
Some systems absolutely require the node name used in an
ssh launch to exactly match that provided in an allocation.
Add the ability to track that name (as opposed to a version
that has the FQDN stripped) and use it in the ssh launcher.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 2482abbd86686f0fa49c01e6da168caa05012e8f)