Enable per-job specification of the exec agent
PRRTE allows the user to specify their own exec agent - i.e.,
the function to use when spawning an application process on
the compute node. In most cases, this is just NULL as we exec
the application process itself. However, in some cases users
want/need a wrapper function that will in turn exec the app.

We have always had an MCA parameter by which one could set that,
but we need to support it in DVM mode as well. So let the MCA
parameter set the default, and add a cmd line option to PRRTE
so one can override the default (whether NULL or some MCA param
setting) for a specific job.

Signed-off-by: Ralph Castain <rhc@pmix.org>