Reset getopt() to 0, not 1.
prrte currently calls parse_cli() twice, once to parse
the schizo command line and once for the app command line.
According to the getopt_long() docs, it should be reset on the
second invocation when parsing the same command line:

"A program that scans multiple argument vectors, or rescans the
 same vector more than once, and wants to make use of GNU
 extensions such as '+' and '-' at the start of optstring, or
 changes the value of POSIXLY_CORRECT between scans, must
 reinitialize getopt() by resetting optind to 0, rather than the
 traditional value of 1.  (Resetting to 0 forces the invocation of
 an internal initialization routine that rechecks POSIXLY_CORRECT
 and checks for GNU extensions in optstring.)"
https://man7.org/linux/man-pages/man3/getopt.3.html

Since prrte makes use of '-' at the start of opt-string - this seems
to apply. I made the mistake of resetting this to 1 to fix a different
bug, instead make sure we don't check for a '-' on the first optind
to ignore prterun/mpirun/prun/ect.

Introduced here: 16d99f20ce35933747dd7d51ab7ecde5c4fbaacb

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>