build: Rewrite libnl checks with OAC_CHECK_PACKAGE
Rewrite the netlink component's configure to check for libnl3
directly, given how simple the libnl checks are given that PRTE
is not using the libnl1 vs. libnl3 checks.

We're not doing the libnl1 vs. libnl3 checks any more (it does not
appear PRTE_CHECK_PACKAGE ever included the hooks to do so), so we
can greatly simplify all the libnl checks.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>