Few small changes to make_dist_tarball
Allow passing of configure args on the cmd line. If --verok is
given, then don't output all the warnings about incorrect tool
versions as it confuses people. Fix a typo in the end - the
tarballs to be moved to dstdir are named "prrte-"

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 7393663b444eb814f3f789dfbb15154efdadcf71)