Fix some PMIx vs PRRTE confusion
The PMIx list of interfaces is being constructed by PMIx, so
PRRTE should definitely not do it again! As the two frameworks
are identical, remove the PRRTE one. Ensure the list gets
initialized prior to PRRTE trying to use it. Update the
various consumers of that list to use the PMIx version.
This allowed prte to at least initialize and run simple
programs across multiple nodes.

However, the "fence" was hanging due to a problem in the
grpcomm_base_stubs.c file. We were calling a pmix_hash_table
function, but comparing the return status to PRRTE values
instead of PMIx values. This caused us to incorrectly decide
that an error had occurred, leading to a hang as the fence
had no way to proceed.

This problem (incorrectly checking return status codes of
the wrong type) is something that probably spans the code
base due to the conversion to using PMIx types. It was
relatively easy to make that conversion - but now we need
to audit the code to ensure every call to a PMIx class
function is properly checked against a PMIx code.

Signed-off-by: Ralph Castain <rhc@pmix.org>