Ensure the proctable includes absolute paths to executable
The requestor has no way of knowing the cwd used for the
executable path if it is given in relative syntax, so use
the appropriate utilities to ensure it is always returned
in absolute path format.

Thanks to @david-edwards-arm for the report AND the
initial patch! Both much appreciated.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit ee9119d1d95a1c0d2eee5de817084d4f5e7d6c62)